refactor: split animation service helpers
