+++ src/core/scene/scene-process/service/animation.ts
@@ -2,20 +2,13 @@ import {
     Animation,
     AnimationClip,
     AnimationState,
-    Component,
     Node,
-    Scene,
-    SkeletalAnimation,
-    animation,
-    assetManager as ccAssetManager,
-    js,
 } from 'cc';
 import {
     AnimationPlayState,
     AnimationEventReason,
     IAnimationClipEvent,
     IAnimationClipDump,
-    IAnimationClipMenuItem,
     IAnimationClipsInfo,
     IAnimationEditClipOptions,
     IAnimationEnterOptions,
@@ -53,100 +46,32 @@ import { queryAuxiliaryCurveValueAtFrame } from './animation/auxiliary-curve';
 import { saveSkeletonAnimationMeta } from './animation/skeleton-meta';
 import { captureAnimationSampledState, restoreAnimationSampledState } from './animation/sampled-state';
 import { AnimationClipSnapshotCommand } from './animation/undo';
-import { IAnimationData, IAnimationSession } from './animation/types';
-import { cloneValue, clipUuid, getClipSample } from './animation/utils';
+import { IAnimationSession } from './animation/types';
+import { clipUuid, getClipSample } from './animation/utils';
 import type { IPropertyCurveMetadataContext } from './animation/property-curve';
 import { queryAnimationPropertyMetadata, queryComponentAnimableProperties } from './animation/property-metadata';
 import { normalizeProvidedAnimationPropertyOperationValue, serializeAnimationPropertyValue } from './animation/property-value';
-
-const NodeMgr = EditorExtends.Node;
-
-const DEFAULT_PROPERTIES: IAnimationPropertyInfo[] = [
-    createPropertyInfo('position', 'cc.Vec3'),
-    createPropertyInfo('eulerAngles', 'cc.Vec3', 'rotation(eulerAngles)'),
-    createPropertyInfo('rotation', 'cc.Quat', 'rotation(quaternion)'),
-    createPropertyInfo('scale', 'cc.Vec3'),
-];
-
-const ACTIVE_PROPERTY = createPropertyInfo('active', 'cc.Boolean');
-
-function createPropertyInfo(name: string, type: string, displayName = name, comp?: string): IAnimationPropertyInfo {
-    return {
-        name,
-        key: comp ? `${comp}.${name}` : name,
-        displayName,
-        type: { value: type },
-        menuName: displayName,
-        comp,
-    };
-}
-
-function isAnimationOperationResult(value: IAnimationOperation | IAnimationOperationResult): value is IAnimationOperationResult {
-    return (value as IAnimationOperationResult).state === 'success' || (value as IAnimationOperationResult).state === 'failure';
-}
-
-function shouldSyncClipDuration(operation: IAnimationOperation): boolean {
-    switch (operation.type) {
-        case 'changeSample':
-        case 'addEvent':
-        case 'deleteEvent':
-        case 'updateEvent':
-        case 'moveEvents':
-        case 'copyEventsTo':
-        case 'addEmbeddedPlayer':
-        case 'deleteEmbeddedPlayer':
-        case 'updateEmbeddedPlayer':
-        case 'clearEmbeddedPlayer':
-        case 'removeEmbeddedPlayerGroup':
-        case 'clearEmbeddedPlayerGroup':
-        case 'removeAuxiliaryCurve':
-        case 'createAuxKey':
-        case 'removeAuxKey':
-        case 'moveAuxKeys':
-        case 'copyAuxKey':
-        case 'createPropertyKey':
-        case 'updatePropertyKey':
-        case 'removePropertyCurve':
-        case 'removePropertyKey':
-        case 'removePropertyKeys':
-        case 'movePropertyKeys':
-        case 'copyPropertyKeysTo':
-            return true;
-        default:
-            return false;
-    }
-}
-
-function isAllowedSkeletonAnimationOperation(operation: IAnimationOperation): boolean {
-    switch (operation.type) {
-        case 'changeSample':
-        case 'changeSpeed':
-        case 'changeWrapMode':
-        case 'addEvent':
-        case 'deleteEvent':
-        case 'updateEvent':
-        case 'moveEvents':
-        case 'copyEventsTo':
-        case 'addEmbeddedPlayer':
-        case 'deleteEmbeddedPlayer':
-        case 'updateEmbeddedPlayer':
-        case 'clearEmbeddedPlayer':
-        case 'addEmbeddedPlayerGroup':
-        case 'removeEmbeddedPlayerGroup':
-        case 'clearEmbeddedPlayerGroup':
-        case 'addAuxiliaryCurve':
-        case 'removeAuxiliaryCurve':
-        case 'renameAuxiliaryCurve':
-        case 'createAuxKey':
-        case 'removeAuxKey':
-        case 'moveAuxKeys':
-        case 'copyAuxKey':
-        case 'updateAuxKeyData':
-            return true;
-        default:
-            return false;
-    }
-}
+import { ACTIVE_PROPERTY, DEFAULT_PROPERTIES } from './animation/property-menu';
+import { isAllowedSkeletonAnimationOperation, isAnimationOperationResult, shouldSyncClipDuration } from './animation/operation-policy';
+import {
+    loadAnimationClip,
+    queryAnimationClipsInfo,
+    queryNodeAnimationData,
+    rebindAnimationComponentClip,
+    resolveAnimationClip,
+} from './animation/clip-library';
+import {
+    extractSampledOperationValue,
+    getAnimationMode,
+    getNodeByPath,
+    getNodeByUuid,
+    getNodePath,
+    isSkeletonClip,
+    isUsingBakedAnimation,
+    queryAnimationComponent,
+    queryAnimationRootNode,
+    readPropertyValue,
+} from './animation/scene-node';
 
 @register('Animation')
 export class AnimationService extends BaseService<Record<string, any>> implements IAnimationService {
@@ -175,9 +100,9 @@ export class AnimationService extends BaseService<Record<string, any>> implement
             await this.exit({ restoreSelection: false, restoreSampledSceneState: true });
         }
 
-        const rootNode = this._queryAnimationRootNode(this._resolveNode(options));
-        const animData = await this._queryNodeAnimationData(rootNode, options.clipUuid, { recoverClipBinding: true });
-        const clip = this._resolveClip(animData, options.clipUuid);
+        const rootNode = queryAnimationRootNode(this._resolveNode(options), Service.Editor.getRootNode());
+        const animData = await queryNodeAnimationData(rootNode, options.clipUuid, { recoverClipBinding: true });
+        const clip = resolveAnimationClip(animData, options.clipUuid);
         const uuid = clipUuid(clip);
         if (!uuid) {
             throw new Error('Animation clip uuid is empty.');
@@ -188,7 +113,7 @@ export class AnimationService extends BaseService<Record<string, any>> implement
             previousSelection: Service.Selection.query(),
             restoreSelectionOnExit: options.restoreSelectionOnExit ?? true,
             rootUuid: rootNode.uuid,
-            rootPath: this._getNodePath(rootNode),
+            rootPath: getNodePath(rootNode),
             clipUuid: uuid,
             sampledRootState: captureAnimationSampledState(rootNode),
         };
@@ -213,7 +138,7 @@ export class AnimationService extends BaseService<Record<string, any>> implement
 
         const shouldRestoreSampledState = options.restoreSampledSceneState ?? true;
         if (shouldRestoreSampledState && session.sampledRootState) {
-            const rootNode = this._getNodeByUuid(session.rootUuid);
+            const rootNode = getNodeByUuid(session.rootUuid);
             if (rootNode) {
                 await restoreAnimationSampledState(rootNode, session.sampledRootState);
                 this._emitNodeChanged(rootNode);
@@ -244,7 +169,7 @@ export class AnimationService extends BaseService<Record<string, any>> implement
             return {
                 active: false,
                 editorType,
-                mode: this._getMode(editorType),
+                mode: getAnimationMode(editorType),
                 rootUuid: '',
                 rootPath: '',
                 clipUuid: '',
@@ -270,16 +195,16 @@ export class AnimationService extends BaseService<Record<string, any>> implement
     }
 
     async queryRoot(options: IAnimationTargetOptions): Promise<IAnimationRootResult> {
-        const rootNode = this._queryAnimationRootNode(this._resolveNode(options));
+        const rootNode = queryAnimationRootNode(this._resolveNode(options), Service.Editor.getRootNode());
         return {
             rootUuid: rootNode.uuid,
-            rootPath: this._getNodePath(rootNode),
+            rootPath: getNodePath(rootNode),
         };
     }
 
     async queryRootInfo(options: IAnimationTargetOptions): Promise<IAnimationRootInfo> {
         const rootNode = this._resolveRootNode(options);
-        const clipsInfo = await this._queryClipsInfo(rootNode);
+        const clipsInfo = await queryAnimationClipsInfo(rootNode);
         const defaultClipUuid = clipsInfo.defaultClip;
         return {
             ...clipsInfo,
@@ -287,12 +212,12 @@ export class AnimationService extends BaseService<Record<string, any>> implement
             clipDump: defaultClipUuid ? await this.queryClip({ rootUuid: rootNode.uuid, clipUuid: defaultClipUuid }) : null,
             time: this._session && defaultClipUuid ? await this.queryTime({ clipUuid: defaultClipUuid }) : 0,
             state: this._playState,
-            useBakedAnimation: this._isUsingBakedAnimation(rootNode),
+            useBakedAnimation: isUsingBakedAnimation(rootNode),
         };
     }
 
     async queryClips(options: IAnimationTargetOptions): Promise<IAnimationClipsInfo> {
-        return await this._queryClipsInfo(this._resolveRootNode(options));
+        return await queryAnimationClipsInfo(this._resolveRootNode(options));
     }
 
     async queryClip(options: IAnimationQueryClipOptions): Promise<IAnimationClipDump> {
@@ -330,7 +255,7 @@ export class AnimationService extends BaseService<Record<string, any>> implement
 
     async queryProperties(options: IAnimationTargetOptions): Promise<IAnimationPropertyInfo[]> {
         const node = this._resolveNode(options);
-        const root = this._session ? this._getNodeByUuid(this._session.rootUuid) : this._queryAnimationRootNode(node);
+        const root = this._session ? getNodeByUuid(this._session.rootUuid) : queryAnimationRootNode(node, Service.Editor.getRootNode());
         const isChild = Boolean(root && root !== node);
         const properties = isChild ? [ACTIVE_PROPERTY, ...DEFAULT_PROPERTIES] : [...DEFAULT_PROPERTIES];
 
@@ -372,7 +297,7 @@ export class AnimationService extends BaseService<Record<string, any>> implement
             state.sample();
 
             const node = this._resolveFrameQueryNode(options);
-            value = this._readPropertyValue(node, options.propKey);
+            value = readPropertyValue(node, options.propKey);
         } finally {
             state.setTime(previousTime);
             if (!state.isPaused) {
@@ -479,7 +404,7 @@ export class AnimationService extends BaseService<Record<string, any>> implement
         }
 
         await this._stopCurrent();
-        this._resolveClip(await this._queryNodeAnimationData(this._getSessionRootNode(), options.clipUuid, { recoverClipBinding: true }), options.clipUuid);
+        resolveAnimationClip(await queryNodeAnimationData(this._getSessionRootNode(), options.clipUuid, { recoverClipBinding: true }), options.clipUuid);
         session.clipUuid = options.clipUuid;
         this._curEditTime = 0;
         await this._getAnimationState(options.clipUuid);
@@ -510,7 +435,7 @@ export class AnimationService extends BaseService<Record<string, any>> implement
                 }
                 return inputFailure;
             }
-            if (this._isSkeletonClip(session.clipUuid, rootNode) && !isAllowedSkeletonAnimationOperation(inputOperation)) {
+            if (isSkeletonClip(session.clipUuid, rootNode) && !isAllowedSkeletonAnimationOperation(inputOperation)) {
                 const skeletonFailure = {
                     state: 'failure',
                     result: false,
@@ -610,8 +535,8 @@ export class AnimationService extends BaseService<Record<string, any>> implement
         const keyData = keyOperation.keyData ?? keyOperation.curveData;
         if (keyOperation.value !== undefined) {
             const value = await normalizeProvidedAnimationPropertyOperationValue(rootNode, rootPath, keyOperation, {
-                queryNodeByUuid: (uuid) => this._getNodeByUuid(uuid),
-                queryNodePath: (node) => this._getNodePath(node),
+                queryNodeByUuid: (uuid) => getNodeByUuid(uuid),
+                queryNodePath: (node) => getNodePath(node),
             });
             return { ...keyOperation, keyData, value };
         }
@@ -631,7 +556,7 @@ export class AnimationService extends BaseService<Record<string, any>> implement
                 propKey: keyOperation.propKey,
                 frame: keyOperation.frame,
             });
-            const value = this._extractSampledOperationValue(sampled, keyOperation.channel);
+            const value = extractSampledOperationValue(sampled, keyOperation.channel);
             if (value === undefined) {
                 return {
                     state: 'failure',
@@ -653,7 +578,7 @@ export class AnimationService extends BaseService<Record<string, any>> implement
     async save(): Promise<boolean> {
         const session = this._requireSession();
         const state = await this._getAnimationState(session.clipUuid);
-        if (this._isSkeletonClip(session.clipUuid, this._getSessionRootNode())) {
+        if (isSkeletonClip(session.clipUuid, this._getSessionRootNode())) {
             await saveSkeletonAnimationMeta(session.clipUuid, state.clip);
             return true;
         }
@@ -698,19 +623,9 @@ export class AnimationService extends BaseService<Record<string, any>> implement
         return this._session;
     }
 
-    private _getMode(editorType: 'scene' | 'prefab' | 'unknown') {
-        if (editorType === 'scene') {
-            return 'general';
-        }
-        if (editorType === 'prefab') {
-            return 'prefab';
-        }
-        return 'unknown';
-    }
-
     private _resolveRootNode(options: IAnimationTargetOptions): Node {
         const node = this._resolveNode(options);
-        return options.rootPath || options.rootUuid ? node : this._queryAnimationRootNode(node);
+        return options.rootPath || options.rootUuid ? node : queryAnimationRootNode(node, Service.Editor.getRootNode());
     }
 
     private _resolveNode(options: IAnimationTargetOptions | IAnimationEnterOptions): Node {
@@ -719,14 +634,14 @@ export class AnimationService extends BaseService<Record<string, any>> implement
         const path = 'rootPath' in options ? options.rootPath : undefined;
         const nodeUuid = 'nodeUuid' in options ? options.nodeUuid : undefined;
         const nodePath = 'nodePath' in options ? options.nodePath : undefined;
-        const target = this._getNodeByUuid(uuid || nodeUuid || '') || this._getNodeByPath(path || nodePath || '');
+        const target = getNodeByUuid(uuid || nodeUuid || '') || getNodeByPath(path || nodePath || '');
         if (target) {
             return target;
         }
 
         const selection = Service.Selection.query();
         if (selection.length > 0) {
-            const selected = this._getNodeByPath(selection[0]);
+            const selected = getNodeByPath(selection[0]);
             if (selected) {
                 return selected;
             }
@@ -737,19 +652,19 @@ export class AnimationService extends BaseService<Record<string, any>> implement
 
     private _resolveFrameQueryNode(options: IAnimationQueryPropertyValueAtFrameOptions): Node {
         const session = this._requireSession();
-        const nodeByUuid = this._getNodeByUuid(options.nodeUuid || '');
+        const nodeByUuid = getNodeByUuid(options.nodeUuid || '');
         if (nodeByUuid) {
             return nodeByUuid;
         }
 
         const path = options.nodePath || session.rootPath;
-        const nodeByPath = this._getNodeByPath(path);
+        const nodeByPath = getNodeByPath(path);
         if (nodeByPath) {
             return nodeByPath;
         }
 
         if (options.nodePath && !options.nodePath.startsWith(session.rootPath)) {
-            const relativeNode = this._getNodeByPath(`${session.rootPath}/${options.nodePath}`);
+            const relativeNode = getNodeByPath(`${session.rootPath}/${options.nodePath}`);
             if (relativeNode) {
                 return relativeNode;
             }
@@ -758,107 +673,22 @@ export class AnimationService extends BaseService<Record<string, any>> implement
         throw new Error(`Animation target node is required: ${path}`);
     }
 
-    private _queryAnimationRootNode(node: Node): Node {
-        let current: Node | null = node;
-        const editorRoot = Service.Editor.getRootNode();
-        while (current) {
-            if (this._queryAnimationComponent(current)) {
-                return current;
-            }
-            if (current === editorRoot || current.parent instanceof Scene) {
-                break;
-            }
-            current = current.parent;
-        }
-        return node;
-    }
-
-    private _queryAnimationComponent(node: Node): Animation | animation.AnimationController | null {
-        const controllerCtor = (animation as any).AnimationController;
-        const controller = controllerCtor ? node.getComponent(controllerCtor) : null;
-        if (controller) {
-            return controller as animation.AnimationController;
-        }
-        return node.getComponent(Animation);
-    }
-
-    private async _queryNodeAnimationData(node: Node, preferredClipUuid?: string, options: { allowEmpty?: boolean; recoverClipBinding?: boolean } = {}): Promise<IAnimationData> {
-        const animComp = this._queryAnimationComponent(node);
-        if (!animComp) {
-            throw new Error(`Animation component not found on node: ${this._getNodePath(node)}`);
-        }
-
-        let clips: AnimationClip[] = [];
-        let defaultClip: AnimationClip | null = null;
-        if (animComp instanceof Animation) {
-            clips = (animComp.clips || []).filter((clip): clip is AnimationClip => Boolean(clip?.name));
-            defaultClip = animComp.defaultClip || clips[0] || null;
-            if (defaultClip?.name) {
-                clips.push(defaultClip);
-            }
-        } else {
-            clips = (await this._visitAnimationClipsInController(animComp))
-                .filter((clip): clip is AnimationClip => Boolean(clip?.name));
-            defaultClip = clips[0] || null;
-        }
-
-        clips = this._uniqAnimationClips(clips);
-        if (!defaultClip?.name) {
-            defaultClip = clips[0] || null;
-        }
-        if (options.recoverClipBinding && (clips.length === 0 || !defaultClip) && preferredClipUuid && animComp instanceof Animation) {
-            const recoveredClip = await this._loadAnimationClip(preferredClipUuid);
-            if (recoveredClip?.name) {
-                this._rebindAnimationComponentClip(animComp, recoveredClip);
-                clips = this._uniqAnimationClips((animComp.clips || []).filter((clip): clip is AnimationClip => Boolean(clip?.name)));
-                defaultClip = animComp.defaultClip?.name ? animComp.defaultClip : clips[0] || null;
-            }
-        }
-        if (clips.length === 0 || !defaultClip) {
-            if (options.allowEmpty) {
-                return { node, animComp, clips, defaultClip };
-            }
-            throw new Error(`Animation clips not found on node: ${this._getNodePath(node)}`);
-        }
-
-        return { node, animComp, clips, defaultClip };
-    }
-
-    private _resolveClip(animData: IAnimationData, uuid?: string): AnimationClip {
-        const targetUuid = uuid || clipUuid(animData.defaultClip);
-        const clip = animData.clips.find((item) => clipUuid(item) === targetUuid);
-        if (!clip) {
-            throw new Error(`Animation clip not found: ${targetUuid}`);
-        }
-        return clip;
-    }
-
-    private async _queryClipsInfo(rootNode: Node): Promise<IAnimationClipsInfo> {
-        const animData = await this._queryNodeAnimationData(rootNode, undefined, { allowEmpty: true });
-        return {
-            rootUuid: rootNode.uuid,
-            rootPath: this._getNodePath(rootNode),
-            clipsMenu: this._decodeClipsMenu(animData.clips),
-            defaultClip: clipUuid(animData.defaultClip),
-        };
-    }
-
     private async _resolveClipForQuery(options: IAnimationQueryClipOptions): Promise<{ rootNode: Node; clip: AnimationClip }> {
         const hasTarget = Boolean(options.rootPath || options.rootUuid || options.nodePath || options.nodeUuid);
         const rootNode = hasTarget ? this._resolveRootNode(options) : this._getSessionRootNode();
         const defaultUuid = this._session?.rootUuid === rootNode.uuid ? this._session.clipUuid : undefined;
         const targetUuid = options.clipUuid || defaultUuid;
-        const animData = await this._queryNodeAnimationData(rootNode, targetUuid);
+        const animData = await queryNodeAnimationData(rootNode, targetUuid);
         return {
             rootNode,
-            clip: this._resolveClip(animData, targetUuid),
+            clip: resolveAnimationClip(animData, targetUuid),
         };
     }
 
     private _createClipDump(rootNode: Node, clip: AnimationClip, state?: AnimationState): IAnimationClipDump {
         return createClipDump(clip, state, {
-            isSkeleton: this._isSkeletonClip(clipUuid(clip), rootNode),
-            useBakedAnimation: this._isUsingBakedAnimation(rootNode),
+            isSkeleton: isSkeletonClip(clipUuid(clip), rootNode),
+            useBakedAnimation: isUsingBakedAnimation(rootNode),
             queryPropertyMetadata: (nodePath, propKey) => queryAnimationPropertyMetadata(rootNode, nodePath, propKey),
         });
     }
@@ -876,7 +706,7 @@ export class AnimationService extends BaseService<Record<string, any>> implement
             return existed;
         }
 
-        const clip = this._resolveClip(await this._queryNodeAnimationData(this._getSessionRootNode(), uuid), uuid);
+        const clip = resolveAnimationClip(await queryNodeAnimationData(this._getSessionRootNode(), uuid), uuid);
         return this._createAnimationState(uuid, clip);
     }
 
@@ -967,53 +797,16 @@ export class AnimationService extends BaseService<Record<string, any>> implement
         }
 
         const time = this._curEditTime;
-        const clip = await this._loadAnimationClip(uuid);
+        const clip = await loadAnimationClip(uuid);
         const rootNode = this._getSessionRootNode();
-        const animComp = this._queryAnimationComponent(rootNode);
+        const animComp = queryAnimationComponent(rootNode);
         if (clip && animComp instanceof Animation) {
-            this._rebindAnimationComponentClip(animComp, clip);
+            rebindAnimationComponentClip(animComp, clip);
         }
         this._resetAnimationState(uuid);
         const state = await this._getAnimationState(uuid);
         await this.setTime({ time: Math.min(time, state.duration) });
         this._broadcastClipChanged('asset-refresh');
-    }
-
-    private _rebindAnimationComponentClip(animComp: Animation, clip: AnimationClip): void {
-        const uuid = clipUuid(clip);
-        const currentDefaultUuid = animComp.defaultClip ? clipUuid(animComp.defaultClip) : '';
-        let found = false;
-        const clips = (animComp.clips || []).map((item) => {
-            if (item && clipUuid(item) === uuid) {
-                found = true;
-                return clip;
-            }
-            return item;
-        });
-        if (!found) {
-            clips.push(clip);
-        }
-        animComp.clips = clips;
-        if (!currentDefaultUuid || currentDefaultUuid === uuid) {
-            animComp.defaultClip = clip;
-        }
-    }
-
-    private async _loadAnimationClip(uuid: string): Promise<AnimationClip | null> {
-        const cached = ccAssetManager.assets.get(uuid);
-        if (cached instanceof AnimationClip) {
-            return cached;
-        }
-
-        return await new Promise((resolve, reject) => {
-            ccAssetManager.loadAny(uuid, (error, asset: AnimationClip) => {
-                if (error) {
-                    reject(error);
-                    return;
-                }
-                resolve(asset instanceof AnimationClip ? asset : null);
-            });
-        });
     }
 
     private _disposeSession(): void {
@@ -1025,59 +818,10 @@ export class AnimationService extends BaseService<Record<string, any>> implement
         this._playState = 'stop';
     }
 
-    private _readPropertyValue(node: Node, propKey: string): unknown {
-        for (const component of node.components) {
-            const names = this._getComponentNames(component);
-            for (const name of names) {
-                const prefix = `${name}.`;
-                if (name && propKey.startsWith(prefix)) {
-                    return this._readPathValue(component, propKey.slice(prefix.length));
-                }
-            }
-        }
-
-        return this._readPathValue(node, propKey);
-    }
-
-    private _extractSampledOperationValue(value: IAnimationValue, channel?: string): IAnimationValue {
-        if (!channel) {
-            return cloneValue(value);
-        }
-        if (!value || typeof value !== 'object' || Array.isArray(value)) {
-            return undefined;
-        }
-        return cloneValue((value as Record<string, IAnimationValue>)[channel]);
-    }
-
-    private _getComponentNames(component: Component): string[] {
-        const names = [
-            js.getClassName(component),
-            (component as any).__className,
-            (component as any).constructor?.__className,
-            (component as any).constructor?.name,
-        ];
-        return names.filter((name, index): name is string => typeof name === 'string' && name.length > 0 && names.indexOf(name) === index);
-    }
-
-    private _readPathValue(target: unknown, path: string): unknown {
-        if (!path) {
-            return target;
-        }
-
-        let value = target as any;
-        for (const key of path.split('.')) {
-            if (value === null || value === undefined) {
-                return undefined;
-            }
-            value = value[key];
-        }
-        return value;
-    }
-
     private _restoreSelection(selection: string[]): void {
         Service.Selection.clear();
         for (const path of selection.slice().reverse()) {
-            if (this._getNodeByPath(path)) {
+            if (getNodeByPath(path)) {
                 Service.Selection.select(path);
             }
         }
@@ -1182,78 +926,11 @@ export class AnimationService extends BaseService<Record<string, any>> implement
 
     private _getSessionRootNode(): Node {
         const session = this._requireSession();
-        const rootNode = this._getNodeByUuid(session.rootUuid);
+        const rootNode = getNodeByUuid(session.rootUuid);
         if (!rootNode) {
             throw new Error(`Animation root node not found: ${session.rootUuid}`);
         }
         return rootNode;
     }
 
-    private _getNodeByUuid(uuid: string): Node | null {
-        if (!uuid) {
-            return null;
-        }
-        return NodeMgr.getNode(uuid) || null;
-    }
-
-    private _getNodeByPath(path: string): Node | null {
-        if (!path) {
-            return null;
-        }
-        return NodeMgr.getNodeByPath(path) || null;
-    }
-
-    private _getNodePath(node: Node): string {
-        return NodeMgr.getNodePath(node) || '';
-    }
-
-    private _decodeClipsMenu(clips: AnimationClip[]): IAnimationClipMenuItem[] {
-        return clips.map((clip) => ({
-            uuid: clipUuid(clip),
-            name: clip.name,
-        }));
-    }
-
-    private _uniqAnimationClips(clips: AnimationClip[]): AnimationClip[] {
-        const seen = new Set<string>();
-        const result: AnimationClip[] = [];
-        for (const clip of clips) {
-            const uuid = clipUuid(clip);
-            if (!uuid || seen.has(uuid)) {
-                continue;
-            }
-            seen.add(uuid);
-            result.push(clip);
-        }
-        return result;
-    }
-
-    private async _visitAnimationClipsInController(controller: animation.AnimationController): Promise<AnimationClip[]> {
-        const system = (globalThis as any).System;
-        if (system?.import) {
-            const mod = await system.import('cc/editor/new-gen-anim');
-            if (typeof mod?.visitAnimationClipsInController === 'function') {
-                return Array.from(mod.visitAnimationClipsInController(controller) as Iterable<AnimationClip>);
-            }
-        }
-
-        const mod = await import('cc/editor/new-gen-anim');
-        if (typeof mod.visitAnimationClipsInController !== 'function') {
-            throw new Error('visitAnimationClipsInController is not available.');
-        }
-        return Array.from(mod.visitAnimationClipsInController(controller) as Iterable<AnimationClip>);
-    }
-
-    private _isUsingBakedAnimation(rootNode: Node): boolean {
-        const animComp = this._queryAnimationComponent(rootNode);
-        return animComp instanceof SkeletalAnimation && Boolean(animComp.useBakedAnimation);
-    }
-
-    private _isSkeletonClip(uuid: string, rootNode?: Node): boolean {
-        if (uuid.includes('@')) {
-            return true;
-        }
-        const targetRootNode = rootNode || (this._session ? this._getNodeByUuid(this._session.rootUuid) : null);
-        return Boolean(targetRootNode && this._queryAnimationComponent(targetRootNode) instanceof SkeletalAnimation);
-    }
 }
